fix: more scalable code and registry
